+++ Sources/UICollectionViewKit/Protocols/ItemDisplayable.swift
@@ -4,9 +4,9 @@ public protocol ItemDisplayable: Hashable {
     var itemID: String { get }
     var categoryID: String { get }
     var imageURL: URL { get }
-    var isAnimatedWebP: Bool { get }
+    var animatedURL: URL? { get }
 }
 
 public extension ItemDisplayable {
-    var isAnimatedWebP: Bool { false }
+    var animatedURL: URL? { nil }
 }